removed the oldConfig completely
