+++ src/lib/global.ts
@@ -2,8 +2,7 @@ import { reactive } from "vue";
 import { getLang, LangType } from "../loc/lang"
 import { Client } from "src/lib/client"
 import { AutoLauncher } from "src/lib/autoLauncher"
-import { oldAppConfig } from "./appConfig";
-import { appConfig } from "./appData"
+import { appConfig } from "./appConfig";
 
 const text: LangType = {}
 // TODO use dependency injection to ensure methods and properties can't be accessed unless they are initialized and valid
@@ -19,7 +18,6 @@ export class Global {
     this.data.loc.text = await getLang(this.data.loc.selected)
   }
   async init(): Promise<void> {
-    oldAppConfig.initAppConfig()
     await appConfig.init()
     await Promise.all(
       [
