+++ src/pages/PlottingProgress.vue
@@ -123,8 +123,7 @@ import { defineComponent } from "vue"
 import { globalState as global } from "src/lib/global"
 import * as util from "src/lib/util"
 import introModal from "components/introModal.vue"
-import { oldAppConfig } from "src/lib/appConfig"
-import { appConfig } from "src/lib/appData"
+import { appConfig } from "src/lib/appConfig"
 
 const lang = global.data.loc.text.plottingProgress
 let farmerTimer: number
@@ -213,23 +212,22 @@ export default defineComponent({
       clearInterval(farmerTimer)
     },
     async farmingWrapper(): Promise<void> {
-      const oldConfig = oldAppConfig.getAppConfig()
       const config = await appConfig.read()
-      if (oldConfig) {
-        await this.client.startBlockSubscription()
-        const farmerStarted = await this.client.startFarming(this.plotDirectory, config.plot.sizeGB)
-        if (!farmerStarted) {
-          console.error("PLOTTING PROGRESS | Farmer start error!")
-        }
-        const networkSegmentCount = oldConfig.segmentCache.networkSegmentCount
-        this.networkSegmentCount = networkSegmentCount
-        this.plottingData.allocatedGB = config.plot.sizeGB
-        this.localSegmentCount = await this.client.getLocalSegmentCount()
-        do {
-          await new Promise((resolve) => setTimeout(resolve, 2000))
-          this.localSegmentCount = await this.client.getLocalSegmentCount()
-        } while (this.localSegmentCount < this.networkSegmentCount)
+
+      await this.client.startBlockSubscription()
+      const farmerStarted = await this.client.startFarming(this.plotDirectory, config.plot.sizeGB)
+      if (!farmerStarted) {
+        console.error("PLOTTING PROGRESS | Farmer start error!")
       }
+      const networkSegmentCount = config.segmentCache.networkSegmentCount
+      this.networkSegmentCount = networkSegmentCount
+      this.plottingData.allocatedGB = config.plot.sizeGB
+      this.localSegmentCount = await this.client.getLocalSegmentCount()
+      do {
+        await new Promise((resolve) => setTimeout(resolve, 2000))
+        this.localSegmentCount = await this.client.getLocalSegmentCount()
+      } while (this.localSegmentCount < this.networkSegmentCount)
+
     },
     startTimers() {
       farmerTimer = window.setInterval(() => {
